# swagger-spec.json
{
  "openapi":"3.0.0",
  "info":{
     "title":"Swagger Document for MVP",
     "description":"API description",
     "version":"1.0",
     "contact":{
        
     }
  },
  "tags":[
     
  ],
  "servers":[
     
  ],
  "components":{
     "schemas":{
        "Email":{
           "type":"object",
           "properties":{
              "id":{
                 "type":"string"
              },
              "name":{
                 "type":"string"
              },
              "description":{
                 "type":"string"
              },
              "type":{
                 "type":"string",
                 "enum":[
                    "PRIMARY",
                    "CC",
                    "BCC"
                 ],
                 "default":"PRIMARY"
              },
              "email":{
                 "type":"string"
              }
           },
           "required":[
              "id",
              "name",
              "description",
              "type",
              "email"
           ]
        },
        "Address":{
           "type":"object",
           "properties":{
              "city":{
                 "type":"string"
              },
              "state":{
                 "type":"string"
              },
              "line1":{
                 "type":"string"
              },
              "line2":{
                 "type":"string"
              },
              "zipCode":{
                 "type":"string"
              }
           },
           "required":[
              "city",
              "state",
              "line1",
              "line2",
              "zipCode"
           ]
        },
        "UserDto":{
           "type":"object",
           "properties":{
              "firstName":{
                 "type":"string"
              },
              "middleName":{
                 "type":"string"
              },
              "lastName":{
                 "type":"string"
              },
              "email":{
                 "$ref":"#/components/schemas/Email"
              },
              "phone":{
                 "type":"number"
              },
              "picture":{
                 "type":"string"
              },
              "address":{
                 "$ref":"#/components/schemas/Address"
              },
              "id":{
                 "type":"string"
              }
           },
           "required":[
              "firstName",
              "middleName",
              "lastName",
              "email",
              "phone",
              "picture",
              "address",
              "id"
           ]
        },
        "RegisterUserDto":{
           "type":"object",
           "properties":{
              "firstName":{
                 "type":"string"
              },
              "middleName":{
                 "type":"string"
              },
              "lastName":{
                 "type":"string"
              },
              "email":{
                 "$ref":"#/components/schemas/Email"
              },
              "phone":{
                 "type":"number"
              },
              "picture":{
                 "type":"string"
              },
              "address":{
                 "$ref":"#/components/schemas/Address"
              }
           },
           "required":[
              "firstName",
              "middleName",
              "lastName",
              "email",
              "phone",
              "picture",
              "address"
           ]
        },
        "LicenseType":{
           "type":"object",
           "properties":{
              "id":{
                 "type":"string"
              },
              "name":{
                 "type":"string"
              },
              "displayName":{
                 "type":"string"
              },
              "description":{
                 "type":"string"
              }
           },
           "required":[
              "id",
              "name",
              "displayName",
              "description"
           ]
        },
        "License":{
            "type":"object",
            "properties":{
               "id":{
                  "type":"string"
               },
               "licenseType":{
                  "$ref":"#/components/schemas/LicenseType"
               },
               "licenseId":{
                  "type":"string"
               },
               "validFromDate":{
                  "type":"string"
               },
               "validToDate":{
                  "type":"string"
               },
               "s3Bucket":{
                  "type":"string"
               },
               "s3Path":{
                  "type":"string"
               },
               "approvedOn":{
                  "type":"string"
               },
               "approvedBy":{
                  "type":"string"
               }
            },
            "required":[
               "id",
               "licenseType",
               "licenseId",
               "validFromDate",
               "validToDate",
               "s3Bucket",
               "s3Path",
               "approvedOn",
               "approvedBy"
            ]
        },
        "Insurance":{
           "type":"object",
           "properties":{
              "id":{
                 "type":"string"
              },
              "insuranceType":{
                 "type":"string",
                 "enum":[
                    "WORKERS_COMPENSATION",
                    "CERTIFICATE_OF_INSURANCE"
                 ],
                 "default":"WORKERS_COMPENSATION"
              },
              "insuranceId":{
                 "type":"string"
              },
              "validFromDate":{
                 "type":"string"
              },
              "validToDate":{
                 "type":"string"
              },
              "s3Bucket":{
                 "type":"string"
              },
              "s3Path":{
                 "type":"string"
              }
           },
           "required":[
               "id",
               "insuranceType",
               "insuranceId",
               "validFromDate",
               "validToDate",
               "s3Bucket",
               "s3Path"
           ]
        },
        "TradeType":{
           "type":"object",
           "properties":{
              "id":{
                 "type":"string"
              },
              "name":{
                 "type":"string"
              },
              "displayName":{
                 "type":"string"
              }
           },
           "required":[
              "id",
              "name",
              "displayName"
           ]
        },
        "Trade":{
           "type":"object",
           "properties":{
              "id":{
                 "type":"string"
              },
              "tradeType":{
                 "$ref":"#/components/schemas/TradeType"
              },
              "approvedOn":{
                 "type":"string"
              },
              "approvedBy":{
                 "type":"string"
              },
              "level":{
                 "type":"number"
              }
           },
           "required":[
              "id",
              "tradeType",
              "approvedOn",
              "approvedBy",
              "level"
           ]
        },
        "Vendor": {
            "type":"object",
            "properties":{
               "licenseList":{
                  "$ref":"#/components/schemas/License"
               },
               "insuranceList":{
                  "$ref":"#/components/schemas/Insurance"
               },
               "tradeList":{
                  "$ref":"#/components/schemas/Trade"
               },
               "companyName":{
                  "type": "string"
               },
               "promoCode":{
                  "type":"string"
               },
               "tosAccepted":{
                  "type":"boolean"
               },
               "payoutEnabled":{
                  "type":"boolean"
               },
               "stripeId":{
                  "type":"string"
               },
               "qbId":{
                  "type":"string"
               }
            },
            "required":[
               "licenseList",
               "insuranceList",
               "tradeList",
               "companyName",
               "promoCode",
               "tosAccepted",
               "payoutEnabled",
               "stripeId",
               "qbId"
            ]
        },
        "Client":{
           "type":"object",
           "properties":{
              "clientType":{
                 "type":"string",
                 "enum":[
                    "HOME_OWNER",
                    "BUYER",
                    "AGENT",
                    "CLOSING_COORDINATOR"
                 ],
                 "default":"HOME_OWNER"
              }
           },
           "required":[
              "clientType"
           ]
        },
        "SuccessResponse":{
           "type":"object",
           "properties":{
              "success":{
                 "type":"boolean"
              }
           },
           "required":[
              "success"
           ]
        }
     }
  },
  "paths":{
     "/api/users":{
        "get":{
           "operationId":"UserController_getUsers",
           "parameters":[
              
           ],
           "responses":{
              "200":{
                 "description":"",
                 "content":{
                    "application/json":{
                       "schema":{
                          "type":"array",
                          "items":{
                             "$ref":"#/components/schemas/UserDto"
                          }
                       }
                    }
                 }
              }
           },
           "tags":[
              "Users Management"
           ],
           "security":[
              {
                 "bearer":[
                    
                 ]
              }
           ]
        },
        "post":{
           "operationId":"UserController_registerUser",
           "parameters":[
              
           ],
           "requestBody":{
              "required":true,
              "content":{
                 "application/json":{
                    "schema":{
                       "$ref":"#/components/schemas/RegisterUserDto"
                    }
                 }
              }
           },
           "responses":{
              "200":{
                 "description":"",
                 "content":{
                    "application/json":{
                       "schema":{
                          "$ref":"#/components/schemas/SuccessResponse"
                       }
                    }
                 }
              }
           },
           "tags":[
              "Users Management"
           ]
        }
     },
     "/api/users/{id}":{
        "get":{
           "operationId":"UserController_getUserByID",
           "parameters":[
              
           ],
           "responses":{
              "200":{
                 "description":"",
                 "content":{
                    "application/json":{
                       "schema":{
                          "$ref":"#/components/schemas/UserDto"
                       }
                    }
                 }
              }
           },
           "tags":[
              "Users Management"
           ],
           "security":[
              {
                 "bearer":[
                    
                 ]
              }
           ]
        },
        "put":{
           "operationId":"UserController_updateUser",
           "parameters":[
              
           ],
           "requestBody":{
              "required":true,
              "content":{
                 "application/json":{
                    "schema":{
                       "$ref":"#/components/schemas/RegisterUserDto"
                    }
                 }
              }
           },
           "responses":{
              "200":{
                 "description":"",
                 "content":{
                    "application/json":{
                       "schema":{
                          "$ref":"#/components/schemas/SuccessResponse"
                       }
                    }
                 }
              }
           },
           "tags":[
              "Users Management"
           ],
           "security":[
              {
                 "bearer":[
                    
                 ]
              }
           ]
        },
        "patch":{
           "operationId":"UserController_patchUser",
           "parameters":[
              
           ],
           "requestBody":{
              "required":true,
              "content":{
                 "application/json":{
                    "schema":{
                       "$ref":"#/components/schemas/RegisterUserDto"
                    }
                 }
              }
           },
           "responses":{
              "200":{
                 "description":"",
                 "content":{
                    "application/json":{
                       "schema":{
                          "$ref":"#/components/schemas/SuccessResponse"
                       }
                    }
                 }
              }
           },
           "tags":[
              "Users Management"
           ],
           "security":[
              {
                 "bearer":[
                    
                 ]
              }
           ]
        },
        "delete":{
           "operationId":"UserController_deleteUser",
           "parameters":[
              
           ],
           "responses":{
              "200":{
                 "description":"",
                 "content":{
                    "application/json":{
                       "schema":{
                          "$ref":"#/components/schemas/SuccessResponse"
                       }
                    }
                 }
              }
           },
           "tags":[
              "Users Management"
           ],
           "security":[
              {
                 "bearer":[
                    
                 ]
              }
           ]
        }
     }
  }
}